Fixing HMD position offset bug in picking & stereo rendering in HMD
